+++ sdk/devhub/arm-devhub/src/classic/workflowOperations/index.ts
@@ -9,86 +9,91 @@ import {
   updateTags,
   createOrUpdate,
   get,
-} from "../../api/workflow/operations.js";
+} from "../../api/workflowOperations/operations.js";
 import type {
-  WorkflowListOptionalParams,
-  WorkflowListByResourceGroupOptionalParams,
-  WorkflowDeleteOptionalParams,
-  WorkflowUpdateTagsOptionalParams,
-  WorkflowCreateOrUpdateOptionalParams,
-  WorkflowGetOptionalParams,
-} from "../../api/workflow/options.js";
+  WorkflowOperationsListOptionalParams,
+  WorkflowOperationsListByResourceGroupOptionalParams,
+  WorkflowOperationsDeleteOptionalParams,
+  WorkflowOperationsUpdateTagsOptionalParams,
+  WorkflowOperationsCreateOrUpdateOptionalParams,
+  WorkflowOperationsGetOptionalParams,
+} from "../../api/workflowOperations/options.js";
 import type { TagsObject, Workflow, DeleteWorkflowResponse } from "../../models/models.js";
 import type { PagedAsyncIterableIterator } from "../../static-helpers/pagingHelpers.js";
 
-/** Interface representing a Workflow operations. */
-export interface WorkflowOperations {
+/** Interface representing a WorkflowOperations operations. */
+export interface WorkflowOperationsOperations {
   /** Gets a list of workflows associated with the specified subscription. */
-  list: (options?: WorkflowListOptionalParams) => PagedAsyncIterableIterator<Workflow>;
+  list: (options?: WorkflowOperationsListOptionalParams) => PagedAsyncIterableIterator<Workflow>;
   /** Gets a list of workflows within a resource group. */
   listByResourceGroup: (
     resourceGroupName: string,
-    options?: WorkflowListByResourceGroupOptionalParams,
+    options?: WorkflowOperationsListByResourceGroupOptionalParams,
   ) => PagedAsyncIterableIterator<Workflow>;
   /** Deletes a workflow */
   delete: (
     resourceGroupName: string,
     workflowName: string,
-    options?: WorkflowDeleteOptionalParams,
+    options?: WorkflowOperationsDeleteOptionalParams,
   ) => Promise<DeleteWorkflowResponse | undefined>;
   /** Updates tags on a workflow. */
   updateTags: (
     resourceGroupName: string,
     workflowName: string,
     parameters: TagsObject,
-    options?: WorkflowUpdateTagsOptionalParams,
+    options?: WorkflowOperationsUpdateTagsOptionalParams,
   ) => Promise<Workflow>;
   /** Creates or updates a workflow */
   createOrUpdate: (
     resourceGroupName: string,
     workflowName: string,
     parameters: Workflow,
-    options?: WorkflowCreateOrUpdateOptionalParams,
+    options?: WorkflowOperationsCreateOrUpdateOptionalParams,
   ) => Promise<Workflow>;
   /** Gets a workflow. */
   get: (
     resourceGroupName: string,
     workflowName: string,
-    options?: WorkflowGetOptionalParams,
+    options?: WorkflowOperationsGetOptionalParams,
   ) => Promise<Workflow>;
 }
 
-function _getWorkflow(context: DeveloperHubServiceContext) {
+function _getWorkflowOperations(context: DeveloperHubServiceContext) {
   return {
-    list: (options?: WorkflowListOptionalParams) => list(context, options),
+    list: (options?: WorkflowOperationsListOptionalParams) => list(context, options),
     listByResourceGroup: (
       resourceGroupName: string,
-      options?: WorkflowListByResourceGroupOptionalParams,
+      options?: WorkflowOperationsListByResourceGroupOptionalParams,
     ) => listByResourceGroup(context, resourceGroupName, options),
     delete: (
       resourceGroupName: string,
       workflowName: string,
-      options?: WorkflowDeleteOptionalParams,
+      options?: WorkflowOperationsDeleteOptionalParams,
     ) => $delete(context, resourceGroupName, workflowName, options),
     updateTags: (
       resourceGroupName: string,
       workflowName: string,
       parameters: TagsObject,
-      options?: WorkflowUpdateTagsOptionalParams,
+      options?: WorkflowOperationsUpdateTagsOptionalParams,
     ) => updateTags(context, resourceGroupName, workflowName, parameters, options),
     createOrUpdate: (
       resourceGroupName: string,
       workflowName: string,
       parameters: Workflow,
-      options?: WorkflowCreateOrUpdateOptionalParams,
+      options?: WorkflowOperationsCreateOrUpdateOptionalParams,
     ) => createOrUpdate(context, resourceGroupName, workflowName, parameters, options),
-    get: (resourceGroupName: string, workflowName: string, options?: WorkflowGetOptionalParams) =>
-      get(context, resourceGroupName, workflowName, options),
+    get: (
+      resourceGroupName: string,
+      workflowName: string,
+      options?: WorkflowOperationsGetOptionalParams,
+    ) => get(context, resourceGroupName, workflowName, options),
   };
 }
 
-export function _getWorkflowOperations(context: DeveloperHubServiceContext): WorkflowOperations {
+export function _getWorkflowOperationsOperations(
+  context: DeveloperHubServiceContext,
+): WorkflowOperationsOperations {
   return {
-    ..._getWorkflow(context),
+    ..._getWorkflowOperations(context),
   };
 }